Added .gitattributes and Lab10 project / .c file
